Faster watershed.
This uses flat indices instead of full blown positions. The access to
the arrays is much faster (in particular cost/status) and the size of
the queue in bytes is much smaller, which should improve our cache
behaviour.

This is circa 5x faster on a ~1M pixel test case (~.5s instead of 2.5s).