ENH Match pymorph erosion convention
The issue is how to handle 0s in the structuring element. In pymorphy,
they are handled as -Infs. It is not clear whether this is the best way
to go about.

Perhaps, signed types should have their minimum value handled as a -Inf,
but it seems odd to do so for unsigned types.