TST Fix weird PNG file
Originally, the file `determinant_zero.png` contained a PNG which
FreeImage was opening incorrectly (it returned the palette instead of
mapping through to the colours).

If correctly mapped, the image was actually full of zeros (i.e., all
colours in the palette were (0,0,0)!)

Therefore, there are actually two tests here:
    with a zero image
    with the values of the palette in determinant_zero.png

The file `determinant_zero.png` now contains what was originally the
palette values.