RFCT Use ndarray_cast to get a data pointer
On development versions, numpy has changed the return type of
PyArray_DATA to be char* instead of void*.
This was reported as issue #26 on github.

This future proofs the code and adds a bunch of assert()s all around to
make sure that we do not have any bad casts. At the cost of an extra
entity, we get a lot of error checking.