RFCT Always use safe form of SAFE_SWITCH_ON...
Removed the non-safe SWITCH_ON... and the last argument to
SAFE_SWITCH_ON_.. It was always true, except for a couple of cases,
which I changed to be true. Then, I removed the argument and made the
behaviour be exactly like the one when true was given in the previous
version.