Update moments.py
When its first order and you are trying to calculate central moments:

\sum_{ij} { img[i,j] (i - c0)**p0 } = dot(img, (p - cm[0])  ** p0)

not p1 as stated in code (I'm still very new to Image Processing so I may be mistaken) 

As proposed, I was able to get First order central moments to be near zero while I wasn't able to do so the way you have it. I can send you the snippet of code i isolated and some sample image runs and their results if you want.